Change i128 back to i32 for performance for now

+++ src/main.rs
@@ -2,7 +2,7 @@ use std::io;
 use rayon::prelude::*;
 use std::fs;
 
-pub fn get_user_input(prompt: &str) -> i128 {
+pub fn get_user_input(prompt: &str) -> i32 {
     println!("{}", prompt);
 
     let mut user_input: String = String::new();
@@ -11,29 +11,29 @@ pub fn get_user_input(prompt: &str) -> i128 {
         .read_line(&mut user_input)
         .expect("Failed to read line");
 
-    let number: i128 = user_input.trim().parse().expect("Please enter an integer!");
+    let number: i32 = user_input.trim().parse().expect("Please enter an integer!");
 
     return number;
 }
 
-fn primes_up_to(number: i128) -> Vec<i128> {
-    (2..=number).into_par_iter().filter(is_prime).collect::<Vec<i128>>()
+fn primes_up_to(number: i32) -> Vec<i32> {
+    (2..=number).into_par_iter().filter(is_prime).collect::<Vec<i32>>()
 }
 
-fn is_prime(number: &i128) -> bool {
+fn is_prime(number: &i32) -> bool {
     if number % 2 == 0 {
         return false;
     }
-    let number_dereferenced: i128 = *number;
-    let root_of_number: i128 = (number_dereferenced as f64).sqrt() as i128;
+    let number_dereferenced: i32 = *number;
+    let root_of_number: i32 = (number_dereferenced as f64).sqrt() as i32;
     !(3..root_of_number).step_by(2).any(|n| root_of_number % n == 0)
 }
 
 fn main() {
-    let user_input: i128 = get_user_input("Enter a positive integer greater than 2: ");
+    let user_input: i32 = get_user_input("Enter a positive integer greater than 2: ");
     
     println!("Calculating primes up to {}...", user_input);
-    let primes: Vec<i128> = primes_up_to(user_input);
+    let primes: Vec<i32> = primes_up_to(user_input);
     
     let prime_str: String = format!("{:?}", primes);
     println!("Saving primes to file...");
